pyproject.toml
[project]
name = "ds-data-driven-stock-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "mpld3>=0.5.12",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.1",
]
